requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
# file was generated using 'pip list --format=freeze > requirements.txt'
alembic==1.13.3
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
basemap==1.4.1
basemap-data==1.3.2
beautifulsoup4==4.12.3
bleach==4.1.0
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
cached-property==1.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorlog==5.0.1
comm==0.2.2
contourpy==1.2.0
cycler==0.11.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
docutils==0.21.2
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
feedgenerator==2.1.0
fonttools==4.51.0
fqdn==1.5.1
graphviz==0.20.1
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.21.0
ipython_genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_contrib_core==0.4.0
jupyter-contrib-nbextensions==0.7.0
jupyter_core==5.7.2
jupyter-events==0.11.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-lsp==2.2.5
jupyter-nbextensions-configurator==0.6.1
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
lxml==5.2.1
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==2.0.4
nb_conda==2.2.1
nb_conda_kernels==2.5.2
nbclient==0.8.0
nbconvert==7.16.5
nbformat==5.10.4
nest_asyncio==1.6.0
notebook==6.4.12
notebook_shim==0.2.4
numexpr==2.10.1
numpy==1.26.4
optuna==4.1.0
optuna-integration==4.1.0
ordered-set==4.1.0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.4
patsy==1.0.1
pelican==4.10.2
pelican-image-process==3.0.4
pelican-jinja-filters==2.1.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.2
platformdirs==3.10.0
plotly==5.24.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.15.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyparsing==3.2.0
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.30.2
requests==2.32.3
rfc3339_validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.10.6
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==75.1.0
six==1.16.0
sklearn-compat==0.1.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.34
stack_data==0.6.3
statsmodels==0.14.4
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.2.1
titlecase==2.4.1
tomli==2.2.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.9.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.11.0
typing_utils==0.1.0
tzdata==2023.3
unicodedata2==15.1.0
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.3.0
watchfiles==1.0.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==4.0.13
xgboost==2.1.1
zipp==3.21.0
zstandard==0.23.0
